Move modular input package to internal
