+++ internal/modularinput/inputs_xml_test.go
