+++ internal/modularinput/testdata/Splunk_TA_otel.inputs.xml
